+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add a scope for included `metadata`.
 - Override metadata when importing `frontmatter` block.
+- Correct the renderer for async.
+- Correct the renderer test that was missing `await`.
+- Add table of content block.
 
 ## [2.0.0] - 2025-08-05